Fix a crash when deleting a row while being moved
Task-number: QT3DS-1755

+++ src/Authoring/Studio/Palettes/TimelineGraphicsView/TimelineGraphicsScene.cpp
@@ -644,7 +644,7 @@ void TimelineGraphicsScene::keyPressEvent(QKeyEvent *keyEvent)
         keyEvent->accept();
         return;
     }
-    if (keyEvent->key() == Qt::Key_Delete)
+    if (keyEvent->key() == Qt::Key_Delete && !m_rowMover->isActive())
         g_StudioApp.DeleteSelectedObject(); // Despite the name, this deletes objects and keyframes
 
     QGraphicsScene::keyPressEvent(keyEvent);
